pyproject.toml
[project]
name = "live-translation"
version = "0.1.0"
description = "Real-time live translator using ADK Gemini Live API"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    "google-adk>=1.28.1",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
